sistema/backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
Pillow>=9.0
django-cors-headers>=4.3